Fix more incorectly named classes / files
